Updated OpenStego to have new property updateStego
The program needs to be able to handle the scenario when the user wants to update the contents of an existing Stego
The updateStego property accepts multiple configuration variables to control the behaviour.
It writes the updated secrets to a new file, backs-up the stego file to a temp holding folder and renames it as a precaution.
After the backup is successful, it creates a new stego file using the newly created secret file and the provided config.
If the creation is unsuccessful, then the backup is restored to the original location and appropriate errors are displayed.
With the successful creation of the new stego, the secret file is deleted to protect the data and a success message is displayed

- Added update stego function to update existing stegos (basically creates a new one but also provides backup processes)
- Added accompanying unit tests to confirm all paths
- Updated the index file to use the update stego function to allow manual testing with real data